+++ iceoryx_binding_c/test/moduletests/test_user_trigger.cpp
@@ -79,14 +79,14 @@ TEST_F(iox_user_trigger_test, cannotBeTriggeredWhenNotAttached)
 
 TEST_F(iox_user_trigger_test, canBeTriggeredWhenAttached)
 {
-    iox_ws_attach_user_trigger_event(&m_waitSet, m_sut, 0U, NULL);
+    iox_ws_attach_user_trigger_event(&m_waitSet, m_sut, 0U, triggerCallback);
     iox_user_trigger_trigger(m_sut);
     EXPECT_TRUE(iox_user_trigger_has_triggered(m_sut));
 }
 
 TEST_F(iox_user_trigger_test, triggeringWaitSetResultsInCorrectNotificationId)
 {
-    iox_ws_attach_user_trigger_event(&m_waitSet, m_sut, 88191U, NULL);
+    iox_ws_attach_user_trigger_event(&m_waitSet, m_sut, 88191U, triggerCallback);
     iox_user_trigger_trigger(m_sut);
 
     auto eventVector = m_waitSet.wait();
@@ -111,9 +111,9 @@ TEST_F(iox_user_trigger_test, triggeringWaitSetResultsInCorrectCallback)
 TEST_F(iox_user_trigger_test, attachingToAnotherWaitSetCleansupFirstWaitset)
 {
     WaitSetMock m_waitSet2{m_condVar};
-    iox_ws_attach_user_trigger_event(&m_waitSet, m_sut, 0U, NULL);
+    iox_ws_attach_user_trigger_event(&m_waitSet, m_sut, 0U, triggerCallback);
 
-    iox_ws_attach_user_trigger_event(&m_waitSet2, m_sut, 0U, NULL);
+    iox_ws_attach_user_trigger_event(&m_waitSet2, m_sut, 0U, triggerCallback);
 
     EXPECT_EQ(m_waitSet.size(), 0U);
     EXPECT_EQ(m_waitSet2.size(), 1U);
@@ -122,7 +122,7 @@ TEST_F(iox_user_trigger_test, attachingToAnotherWaitSetCleansupFirstWaitset)
 TEST_F(iox_user_trigger_test, disable_trigger_eventingItFromWaitsetCleansup)
 {
     WaitSetMock m_waitSet2{m_condVar};
-    iox_ws_attach_user_trigger_event(&m_waitSet, m_sut, 0U, NULL);
+    iox_ws_attach_user_trigger_event(&m_waitSet, m_sut, 0U, triggerCallback);
 
     iox_ws_detach_user_trigger_event(&m_waitSet, m_sut);