+++ tools/ci/build-test-ubuntu-with-sanitizers.sh
@@ -19,6 +19,8 @@
 
 set -e
 
+COMPILER=${1:-gcc}
+
 msg() {
     printf "\033[1;32m%s: %s\033[0m\n" ${FUNCNAME[1]} "$1"
 }
@@ -29,15 +31,29 @@ cd ${WORKSPACE}
 msg "installing build dependencies"
 sudo apt-get update && sudo apt-get install -y libacl1-dev libncurses5-dev
 
+if [ "$COMPILER" == "clang" ]; then
+    msg "installing latest stable clang"
+    wget https://apt.llvm.org/llvm.sh -O /tmp/llvm.sh
+    chmod +x /tmp/llvm.sh
+    # set LLVM_VERSION
+    eval $(cat /tmp/llvm.sh | grep LLVM_VERSION= -m 1)
+    sudo /tmp/llvm.sh ${LLVM_VERSION}
+fi
+
 msg "creating local test users and groups for testing access control"
 sudo ./tools/add_test_users.sh
 
-msg "compiler versions:
-$(gcc --version)
-$(clang --version)"
 
 msg "building sources"
-./tools/iceoryx_build_test.sh clean build-strict build-strict build-all clang sanitize test-add-user
+if [ "$COMPILER" == "gcc" ]; then
+    ./tools/iceoryx_build_test.sh clean build-strict build-strict build-all debug sanitize test-add-user
+fi
+
+if [ "$COMPILER" == "clang" ]; then
+    export CC=clang-${LLVM_VERSION}
+    export CXX=clang++-${LLVM_VERSION}
+    ./tools/iceoryx_build_test.sh clean build-strict build-strict build-all clang debug sanitize test-add-user
+fi
 
 msg "running all tests"
 cd ./build